Chore: 커스텀통계 추가를 위한 리펙토링
